GUI: components moved in a separate file
